server/requirements.txt
grpcio==1.66.2
grpcio-tools==1.66.2
protobuf==5.28.2
ipwhois==1.3.0
psutil==7.0.0
dotenv==0.9.9
einops==0.6.0
timm==0.6.12
imutils==0.5.4
opencv-python==4.6.0.66
numpy==1.23.4 
torchvision==0.14.0 